actualizacion de los roles y mejoras de seguridad en la pagina

+++ app/Http/Controllers/RoleController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Role;
 use Spatie\Permission\Models\Permission;
@@ -9,15 +10,17 @@ use Spatie\Permission\Models\Permission;
 
 class RoleController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
     public function index(Request $request)
     {
+        $this->authorize('role-list');
         $texto = $request->input('texto'); //Variable del texto de búsqueda
         $registros=Role::with('permissions')->where('name', 'like', "%{$texto}")
                             ->orderBy('id', 'desc')
-                            ->paginate(2);
+                            ->paginate(10);
         return view('role.index', compact('registros', 'texto'));
     }
 
@@ -26,7 +29,9 @@ class RoleController extends Controller
      */
     public function create()
     {
+        $this->authorize('role-create');
         $permissions = Permission::all();
+
         return view('role.action', compact('permissions'));
     }
 
@@ -35,6 +40,7 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('role-store');
         $request->validate([
             'name' => 'required|unique:roles,name',
             'permissions' => 'required|array',
@@ -59,6 +65,7 @@ class RoleController extends Controller
      */
     public function edit(string $id)
     {
+        $this->authorize('role-edit');
         $registro = Role::findOrFail($id);
         $permissions = Permission::all();
 
@@ -70,6 +77,7 @@ class RoleController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $this->authorize('role-update');
         $registro = Role::findOrFail($id);
         $request->validate([
             'name' => 'required|unique:roles,name,'.$registro->id,
@@ -78,6 +86,7 @@ class RoleController extends Controller
 
         $registro->update(['name' => $request->name]);
         $registro->syncPermissions($request->permissions);
+        return redirect()->route('roles.index')->with('mensaje', 'Registro '.$registro->name.' actualizado exitosamente.');
     }
 
     /**
@@ -85,6 +94,7 @@ class RoleController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('role-delete');
         $registro = Role::findOrFail($id);
         $registro->delete();
